Remove AutoID class
We no longer need to keep creatures of a certain type in a specific id
range as the creature type now is sent when adding a creature in the
protocol. Nor is it a critical section when incrementing the ID counter
because creatures are either created during loading in the main thread,
or after loading in the dispatcher while the server is running.