Remove SimOne's XML npc system
The XML system was too limited, and I haven't seen anyone actually use
it. It didn't seem to be working either with the example Oracle script,
so I've replaced the default Oracle script with one scripted using the
Lua NPC system.

The Lua system is much more extensible, and the idea with 1.0 is to make
the core as small as possible and use Lua to implement features.