Move skills into players table
There's no advantage to storing skills in a separate table if every
player has a row for each skill anyway. It would make more sense if we
didn't insert the rows with value 10 and count 0, but that introduces
another set of problems. This will save 7 queries when saving players, 1
query when loading players, 7 queries when inserting new players, and
removes a trigger (hopefully we can get rid of all triggers soon,
they're bad because MySQL requires SUPER privilege to create triggers
and doesn't warn the user if it didn't insert them for that reason).