Remove md5 password hashing
SHA1 should be replaced too (or at least be strengthened with salt, if
that is even considered enough with SHA1 today), but with the opinions
at hand (plain, MD5 and SHA1) there was never any good reason to choose
MD5 over SHA1.

If you were using MD5, consider resetting the passwords of all your
accounts; or if you really want to continue using MD5, you'll have to
add it back in your own source code copy.