Remove plaintext password type
Due to a bug in server_config "encryption", it's not possible for the
migration script to automatically hash the passwords. If you were using
plain-text passwords, run this query to hash your passwords:
UPDATE `accounts` SET `password` = SHA1(`password`)

This commit also removes critical hit chance from weapons, such
customization is meant to be scripted instead of being part of the core.