Added a prototype version of a casting system
Added a casting system which does all the broadcasting work in the ASIO
thread. Changed ProtocolSpectator::parsePacket() so that
sendCancelWalk() is sent whenever the spectator tries to move/change
direction. Changed disconnectClient() to disconnect() in
ProtocolGame::stopLiveCast() to allow the client to clear its creature
cache. Check if is a known creature to prevent debug assertions on tibia
client. Moved methods and variables that are common to the game and
spectator protocols to an abstract base class - ProtocolGameBase.
The caster will now be able to cast spells and give commands
in the cast chat. Moved the livecast count check outside
the critical section. Live casts will now be properly removed
from the cast list. Added a missing move packet in
ProtocolSpectator::parsePacket() and removed unused ones.