Actions using Metatable everywhere
Not complete yet, but someone can continue the work.
Renamed ItemEx to targetEx since ItemEx can be either a Item or a
Creature so the name make more sense.
The scripts still need need to check if it is a creature or a item that
what remains to be done.