Major cleanup/refactoring
This commit breaks backwards compatibility with breakchance and
ammoaction in items.xml and also certain global event names. Break
chance and action for weapons are now specified in weapons.xml and can
be applied to any weapon. This commit also removes the config option
"removeAmmoWhenUsingDistanceWeapon" as there were issues with that and
it is preferred that end-users configure actions for their weapons
(which provides better customizability) instead of having a global
variable that toggles it for every weapon.

Closes #1125