Remove PositionEx
This commit removes stackpos from position tables sent to onUse and
onCastSpell (in variant position). This was done because calculating the
stackpos every call is not worth it when it's rarely used, it's much
more efficient to manually get the stackpos using Tile.getThingIndex
when it's needed.

This commit also removes a few old functions from the NPC Lua interface
that haven't been in use for a long time: selfGetPosition(), getNpcPos()
and getNpcName().