Minor spells update
Added missing secondary group "special" (and their respective cooldowns)
for the spells:
* Strong Ice Strike
* Strong Terra Strike
* Strong Energy Strike
* Strong Flame Ftrike
* Lightning

Fixed the "range" for the spell "Lightning" - the correct range is 5.
Modified "Condition(CONDITION_HASTE)" to "Condition(CONDITION_PARALYZE)"
in spell "Sharpshooter" - so the "paralyze icon" shows up
(so the player can easily acknowledge he's been paralyzed).
Renamed "AREA_SHORTWAVE3" to "AREA_WAVE3" in spell "Strong Ice Wave".
Renamed "exhaust" to "cooldown" for clarity and because the value in
exhaust stands for the spell/rune cooldown.