Fix bug with normal_distribution
normal_random would return the min or max values with higher probability
than min+1 or max+1 with wide ranges as normal_distribution could return
values below 0.0 and above 1.0. Instead, we now return the middle values
when we get a value below 0.0 or above 1.0, as it gives a more normal
distribution that way. Pull requests welcome if anyone knows a better
way to address this.