Move routing to lua
Routing is now done purely in lua. Switched timer to deadline timer and
changed the timeout to 30 minutes to still be able to detect connection
drops and support long-polling. Added the getRequestMethod function to
the Responder metatable. Router no longer uses the Events API. Removed
the TimerType enum because it's no longer needed (due to changes to
support long-polling).

Signed-off-by: Damian Jarek <damian.jarek93@gmail.com>