create_item.lua bugfix
Fixed bugs if wrong name, a blank item would be created. Also if param left blank. (This is because ItemType returns a blank item with id 8893 if given a nil value). Creation of such items can lead to things and creatures disappearing on that tile.

Another one of my concerns is creating items with ID's such as e.g. 1, which returns the name "water" but crashes the client when created.

I kind of consider this a tempfix though. I suppose this should be solved in the sources.

	// ItemType(id or name)
	uint32_t id;
	if (isNumber(L, 2)) {
		id = getNumber<uint32_t>(L, 2);
	} else {
		id = Item::items.getItemIdByName(getString(L, 2));
	}

Why do you expect it to be a string if its not a number? Should use isString, else pushnil?

(how to format the code?)