IPv6 support
Implemented support for IPv6 connections. The server now listens on an
IPv6 socket unless it's configured to only bind to a specific IPv4
address. The address is now stored as a varchar in the database.
The internal representation of the address has been changed to the
asio-native one for simplicity.

Signed-off-by: Damian Jarek <damian.jarek93@gmail.com>