Added ignoreendvalue to the example quest (#2514)
Setting `ignoreendvalue` will cause the mission to show up in the quest log if the player's storage value is greater than or equal to the start value. If the storage value is greater than the end value it will still show the mission (the last description of the mission), whereas without `ignoreendvalue` the mission would disappear again.