Save storages in lib
In my understanding I think it would be interesting for the repository to start using a lib to store the storages, because in a way it makes reading easier and there is also a great improvement in the organization of the files.

This is my opinion of course, but I want it to be discussed before attaching it.