Define coins ( items.xml )
With this changes in the sources, we can now add new coins in the game without much problem, this feature was on the old servers, but I don't know why it was removed.

The `Game::addMoney` function is not mine, I tried it and it works very well, who agrees with this change?

I know that tfs is based on emulating at all costs Tibia RL, but this change does not affect that and gives the opportunity to people who use this base for their custom projects.