Upgrade raids to lua under globalevents
Hello guys.
I was thinking... The raid system being separated from globalevents is kind of redundant. Globalevents does the same, and it's not xml so you can do better, more detailed scripts for certain raids. I've included this script I made, but this is just an example, since it lacks delayed raids or multiple stages raids. So the script needs to be upgraded. Keep in mind that my script is designed run with a chance of no raids occurring, so an interval of 30 minutes is advised.
It's also a decision on if we should remove the raid system we currently use, so let's discuss this, please.