Correct misuses of __has_cpp_attribute
`__has_cpp_attribute` is for detecting support for attribute-tokens like `fallthrough` or `unlikely`. This queries with feature-test macro names are always going to fail: no implementation will ever support an attribute-token with the same name as a feature-test macro. Given the conditional use of library facilities, the intent here was clearly to test the feature-test macros to determine support for those facilities, so let's do that instead.

I've also fixed the using-declaration for `std::to_underlying`.

Drive-by: Removed extraneous `inline` in the definition of the `constexpr` `to_underlying` fallback; `constexpr` functions are implicitly inline.