Fix memory leak in spawns
When we spawn a monster, we increment the count on the monster object for the entry into the spawnedmap. When we remove the monster from the spawnedmap because of it being outside the zone, we don't decrement the ref count, which leads to a leak.

There is also extra logic in place for a spawn ID 0. Not sure why monsters were being removed from the spawned map and placed back into it with 0 as the id for when they leave the area, but I find no other interactions or checks for spawn id being 0, so there seems to be zero reason, and whatever was so special about spawn id 0 doesn't exist anymore, so I cleaned up that old code that has not been in use since apparently before TFS was TFS.