Fix: Unable to set spell group to 'NONE'
Using the spellgroup("none"), didn't work as below in the code, after it was already loaded, the code: if (group == SPELLGROUP_NONE) {, would have overriden it and set none to either attack or healing group.

This successfully fixes it.

If group is none, then we don't want to load/change group cooldowns, as other checks in the code for spell cooldown check only groupCooldown > 0, and not group != SPELLGROUP_NONE, so I think it will be more efficient to block it at loading.