refactor(fileloader): iterate OTB with direct pointer access (#2)
Improves file loader performance by having less indirection and skipping
checks aggressively.

When running a release build, most of the sanity checks will be skipped
(e.g. checking if the file does not underflow, if the nodes have
expected types) since we expect map editors to generate correct map
files. The map only needs to be loaded once with a debug build if one
wants to make sure it works, so there is no need to check again on every
start.