chore(vs build): update for VS2026 v14.50 (#22)
### Pull Request Prelude
- [x] I have followed [proper The Forgotten Server code styling][code].
- [x] I have read and understood the [contribution guidelines][cont]
before making this PR.
- [x] I am aware that this PR may be closed if the above-mentioned
criteria are not fulfilled.

### Changes Proposed
With the upgrade to the latest [Visual Studio
release](https://devblogs.microsoft.com/visualstudio/visual-studio-2026-is-here-faster-smarter-and-a-hit-with-early-adopters/),
now targeting VS2026 and the v14.50 toolset, we were able to modernize
the entire project build config. This update improves compatibility with
the newest compiler features, streamlines integration with the current
toolchain and ensures a more consistent structure for long-term
maintenance.

Additionally, the previously used `.props` files were removed and their
settings were migrated directly into the `.vcxproj` to reduce
indirection and ensure clearer, centralized configuration across all
build profiles.

[cont]: https://github.com/otland/forgottenserver/wiki/Contributing
[code]:
https://github.com/otland/forgottenserver/wiki/TFS-Coding-Style-Guide