fix(build): reorder CMake statements to enable IPO (#5)
This pull request fixes the build configuration to _actually_ enable
IPO, which was missing since the target is created before
`CMAKE_INTERPROCEDURAL_OPTIMIZATION` is set thus not inheriting it's
value.

Also, since the helper `main.cpp` file is not needed anymore (it was a
work around CMake outputting the binary to `build/src/tfs` instead of
`build/tfs`), inlined that file back into `otserv.cpp`, and dropped the
header.

**Issues addressed:** IPO was never applied to builds.

**How to test:** build with `--verbose` and check if `-flto` is added to
the command.

---------

Co-authored-by: Ramon Bernardo <ramonbernardo.contato@gmail.com>